Fruit-Classifier-Robot-Arm-main/My_Robotic_arm_project/requirements.txt
opencv_python==4.7.0.72
Pillow==9.4.0
Pillow==9.5.0
pyserial==3.5
roboflow==1.0.9
